Refactor: Create top-level config component
