+++ ui/src/app/routing/app-routing.tsx
@@ -28,7 +28,6 @@ import {
 import { bindApiClients as notebooksBindApiClients } from 'app/services/notebooks-swagger-fetch-clients';
 import {
   bindApiClients,
-  configApi,
   getApiBaseUrl,
 } from 'app/services/swagger-fetch-clients';
 import {
@@ -47,7 +46,6 @@ import {
   authStore,
   serverConfigStore,
   stackdriverErrorReporterStore,
-  useStore,
 } from 'app/utils/stores';
 import { StackdriverErrorReporter } from 'stackdriver-errors-js';
 
@@ -111,21 +109,6 @@ const ScrollToTop = () => {
   return <React.Fragment />;
 };
 
-const useServerConfig = () => {
-  const { config } = useStore(serverConfigStore);
-
-  useEffect(() => {
-    const load = async () => {
-      const serverConfig = await configApi().getConfig();
-      serverConfigStore.set({ config: serverConfig });
-    };
-
-    load();
-  }, []);
-
-  return config;
-};
-
 const useOverriddenApiUrl = () => {
   const [overriddenUrl, setOverriddenUrl] = useState('');
 
@@ -166,7 +149,6 @@ const useOverriddenApiUrl = () => {
 };
 
 export const AppRoutingComponent: React.FunctionComponent = () => {
-  const config = useServerConfig();
   const { authLoaded, authError } = useAuthentication();
   const isUserDisabledInDb = useIsUserDisabled();
   const overriddenUrl = useOverriddenApiUrl();
@@ -174,11 +156,8 @@ export const AppRoutingComponent: React.FunctionComponent = () => {
   const loadLocalStorageAccessToken = () => {
     // Ordinarily this sort of thing would go in authentication.tsx - but setting authStore in there causes
     // an infinite loop
-    // Enable test access token override via local storage. Intended to support
-    // Puppeteer testing flows. This is handled in the server config callback
-    // for signin timing consistency. Normally we cannot sign in until we've
-    // loaded the oauth client ID from the config service.
-    if (config && environment.allowTestAccessTokenOverride && !authLoaded) {
+    // Enable test access token override via local storage. Intended to support Puppeteer testing flows.
+    if (environment.allowTestAccessTokenOverride && !authLoaded) {
       const localStorageTestAccessToken = window.localStorage.getItem(
         LOCAL_STORAGE_KEY_TEST_ACCESS_TOKEN
       );
@@ -195,14 +174,11 @@ export const AppRoutingComponent: React.FunctionComponent = () => {
   const redirectToTOSPage = useShowTOS();
 
   useEffect(() => {
-    if (config) {
-      // Bootstrapping that requires server config
-      bindClients();
-      loadErrorReporter();
-      initializeAnalytics();
-      loadLocalStorageAccessToken();
-    }
-  }, [config]);
+    bindClients();
+    loadErrorReporter();
+    initializeAnalytics();
+    loadLocalStorageAccessToken();
+  }, []);
 
   const firstPartyCookiesEnabled = cookiesEnabled();
   const thirdPartyCookiesEnabled = !(
@@ -213,10 +189,6 @@ export const AppRoutingComponent: React.FunctionComponent = () => {
 
   return (
     <React.Fragment>
-      {/* for outdated-browser-rework https://www.npmjs.com/package/outdated-browser-rework*/}
-      {/* Check checkBrowserSupport() function defined in index.ts and implemented in setup.ts*/}
-      {/* The outdated browser banner should be shown on all pages not just for authenticated user*/}
-      <div id='outdated' />
       {authLoaded && isUserDisabledInDb !== undefined && (
         <React.Fragment>
           {/* Once Angular is removed the app structure will change and we can put this in a more appropriate place */}
